Extend diffusion-limited rate calculation to reactions of higher orders
As derived in Flegg (2016) in SIAM J. Appl. Math., the well-known
Smoluchowski diffusion-limited reaction kinetics can be extended to
reactions of any order. Theoretically, the Smoluchowski radius is a
parameter determined through experiment, but in lieu of data, use the
sum of molecular radii as determined from the McGowan volume.